[preview card] Refactor to `useRenderElement`
